+++ pyaptamer/__init__.py
@@ -2,4 +2,8 @@
 
 from importlib.metadata import version
 
-__version__ = version("pyaptamer")
+try:
+    __version__ = version("pyaptamer")
+except Exception:
+    __version__ = "0.0.1-dev"
+
