+++ pyaptamer/trafos/pseaac/__init__.py
@@ -1,0 +1,7 @@
+"""Protein sequence transformers based on PSeAAC."""
+
+__author__ = ["nennomp", "satvshr", "fkiraly"]
+__all__ = ["AptaNetPSeAAC", "PSeAAC"]
+
+from pyaptamer.trafos.pseaac._pseaac_aptanet import AptaNetPSeAAC
+from pyaptamer.trafos.pseaac._pseaac_general import PSeAAC
